requirement.txt
fastapi
uvicorn
pillow
numpy
tensorflow-cpu
runpod
